test(Checkbox): update test cases
